perf(core): colocate functions with Tokyo DB + kill redundant auth round trips
The root cause of app-wide slowness: Supabase lives in ap-northeast-1 (Tokyo)
but all Vercel functions ran in iad1 (US East) — every DB/GoTrue round trip
paid ~170ms cross-Pacific, and a /dashboard load makes ~4 auth calls + ~12
queries over 6-7 sequential waves. This PR removes the distance and most of
the redundant calls. No feature/route change (route table stays 152).

- vercel.json: "regions": ["hnd1"] — functions now run next to the DB, cutting
  per-round-trip latency from ~170ms to ~2ms. Stripe/Resend calls are on
  async-tolerant paths; all 13 crons are DB-bound and get faster.
- middleware: public/self-authenticating routes (portal, booking, forms,
  cron, mcp, oauth, tokenised PDFs, bearer) now short-circuit BEFORE any auth
  call — they never touch GoTrue. Protected routes switch getUser() →
  getClaims(): local JWT verification via cached JWKS (zero network on
  asymmetric signing keys; graceful network fallback on legacy keys). Session
  refresh still happens inside getClaims; RLS remains the security gate.
- auth.ts: getUser/getUserOrNull now share ONE cache() memo (they were two
  separate memos = two GoTrue calls per request from layout + page).
- New src/lib/role.ts getMyRoleCached(): React-cache()d role resolution with
  the owner/member lookups in parallel; my-role.ts delegates ("use server"
  files can't export cache()-wrapped values).
- dashboard/page.tsx: dropped its raw supabase.auth.getUser() (a 4th GoTrue
  call) and its duplicated role-resolution queries; uses the shared helpers.

Verified: tsc clean, 17/17 tests, route count 152.
