+++ src/app/(dashboard)/dashboard/page.tsx
@@ -1,7 +1,7 @@
 import { redirect } from "next/navigation";
-import { createClient } from "@/lib/supabase/server";
-import { getActiveBizId } from "@/lib/active-business";
-import { isWorker, type Role } from "@/lib/permissions";
+import { getUserOrNull } from "@/lib/auth";
+import { getMyRoleCached } from "@/lib/role";
+import { isWorker } from "@/lib/permissions";
 import { getDashboardStats } from "@/lib/actions/invoices";
 import { getBusiness } from "@/lib/actions/business";
 import { getTodayWorkOrders, getWorkOrders } from "@/lib/actions/work-orders";
@@ -9,27 +9,12 @@ import { DashboardClient } from "@/components/dashboard/dashboard-client";
 import { WorkerDashboard } from "@/components/dashboard/worker-dashboard";
 
 export default async function DashboardPage() {
-  const supabase = await createClient();
-  const { data: { user } } = await supabase.auth.getUser();
+  // getUserOrNull shares the layout's cached GoTrue call; getMyRoleCached
+  // shares role resolution instead of re-querying businesses/members here.
+  const user = await getUserOrNull();
   if (!user) redirect("/auth/login");
 
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const sb = supabase as any;
-  const businessId = await getActiveBizId(sb, user.id);
-
-  // Resolve role
-  const { data: biz } = await sb.from("businesses").select("user_id").eq("id", businessId).single();
-  let role: Role = "owner";
-  if (biz?.user_id !== user.id) {
-    const { data: m } = await sb
-      .from("business_members")
-      .select("role")
-      .eq("business_id", businessId)
-      .eq("user_id", user.id)
-      .eq("status", "active")
-      .maybeSingle();
-    role = (m?.role ?? "viewer") as Role;
-  }
+  const role = await getMyRoleCached();
 
   if (isWorker(role)) {
     const [business, todayJobs, allJobs] = await Promise.all([
